+++ src/gdconvert/conversion_ints.cpp
@@ -1,11 +1,11 @@
 #include "conversion_ints.h"
 
-template <typename C, typename T>
-T variant_as_ints_(const Variant& variant) {
+template <typename C>
+C variant_as_int_strict(const Variant &variant) {
     switch (variant.get_type()) {
         case Variant::OBJECT: {
             if (const auto ndarray = Object::cast_to<NDArray>(variant)) {
-                return std::visit([](const auto& carray) -> T {
+                return std::visit([](const auto &carray) -> C {
                     using V = typename std::decay_t<decltype(carray)>::value_type;
 
                     if constexpr (!std::is_integral_v<V>) {
@@ -14,7 +14,37 @@ T variant_as_ints_(const Variant& variant) {
 
                     switch (carray.dimension()) {
                         case 0:
-                            return T { C(*carray.data()) };
+                            return static_cast<C>(*carray.data());
+                        default:
+                            throw std::runtime_error("array must be zero-dimensional or one-dimensional");
+                    }
+                }, ndarray->array.to_compute_variant());
+            };
+        }
+        case Variant::INT:
+            return static_cast<C>(static_cast<int64_t>(variant));
+        default:
+            break;
+    }
+
+    throw std::runtime_error("Variant cannot be converted to an axis type.");
+}
+
+template<typename C, typename T>
+T variant_as_ints_(const Variant &variant) {
+    switch (variant.get_type()) {
+        case Variant::OBJECT: {
+            if (const auto ndarray = Object::cast_to<NDArray>(variant)) {
+                return std::visit([](const auto &carray) -> T {
+                    using V = typename std::decay_t<decltype(carray)>::value_type;
+
+                    if constexpr (!std::is_integral_v<V>) {
+                        throw std::runtime_error("incompatible dtype; must be int");
+                    }
+
+                    switch (carray.dimension()) {
+                        case 0:
+                            return T{C(*carray.data())};
                         case 1: {
                             T ints;
                             ints.resize(carray.size());
@@ -30,17 +60,14 @@ T variant_as_ints_(const Variant& variant) {
         }
         case Variant::ARRAY: {
             const Array axes_array = variant;
-            auto values = T(axes_array.size());
+            auto axes = T(axes_array.size());
             for (int64_t i = 0; i < axes_array.size(); i++) {
-                const Variant& element = axes_array[i];
-                if (element.get_type() != Variant::INT)
-                    throw std::runtime_error("Axis must be an integer");
-                values[i] = static_cast<int64_t>(element);
+                axes[i] = variant_as_int_strict<C>(axes_array[i]);
             }
-            return values;
+            return axes;
         }
         case Variant::INT:
-            return { C(static_cast<int64_t>(variant)) };
+            return {C(static_cast<int64_t>(variant))};
         case Variant::PACKED_BYTE_ARRAY:
             return packed_as_array<T>(PackedByteArray(variant));
         case Variant::PACKED_INT32_ARRAY:
@@ -49,15 +76,15 @@ T variant_as_ints_(const Variant& variant) {
             return packed_as_array<T>(PackedInt64Array(variant));
         case Variant::VECTOR2I: {
             auto vector = Vector2i(variant);
-            return { C(vector.x), C(vector.y) };
+            return {C(vector.x), C(vector.y)};
         }
         case Variant::VECTOR3I: {
             auto vector = Vector3i(variant);
-            return { C(vector.x), C(vector.y), C(vector.z) };
+            return {C(vector.x), C(vector.y), C(vector.z)};
         }
         case Variant::VECTOR4I: {
             auto vector = Vector4i(variant);
-            return { C(vector.x), C(vector.y), C(vector.z), C(vector.w) };
+            return {C(vector.x), C(vector.y), C(vector.z), C(vector.w)};
         }
         default:
             break;
@@ -67,9 +94,17 @@ T variant_as_ints_(const Variant& variant) {
 }
 
 va::shape_type variant_to_shape(const Variant &variant) {
-    return variant_as_ints_<size_t, std::vector<size_t>>(variant);
+    return variant_as_ints_<size_t, std::vector<size_t> >(variant);
 }
 
 va::strides_type variant_to_axes(const Variant &variant) {
     return variant_as_ints_<std::ptrdiff_t, va::strides_type>(variant);
+}
+
+va::axes_type variants_to_axes(const Variant **args, GDExtensionInt arg_count, GDExtensionCallError &error) {
+    auto axes = va::axes_type(arg_count);
+    for (int i = 0; i < arg_count; i++) {
+        axes[i] = variant_as_int_strict<std::ptrdiff_t>(*args[i]);
+    }
+    return axes;
 }