+++ src/gdconvert/conversion_ints.h
@@ -19,5 +19,6 @@ T packed_as_array(Packed packed) {
 
 va::shape_type variant_to_shape(const Variant &variant);
 va::axes_type variant_to_axes(const Variant &variant);
+va::axes_type variants_to_axes(const Variant **args, GDExtensionInt arg_count, GDExtensionCallError &error);
 
 #endif