+++ src/vatensor/linalg.h
@@ -5,7 +5,7 @@
 #include "varray.h"
 
 namespace va {
-    VConstant reduce_dot(const VArray &a, const VArray &b);
+    VScalar reduce_dot(const VArray &a, const VArray &b);
     void reduce_dot(VArrayTarget target, const VArray &a, const VArray &b, const axes_type& axes);
 
     void dot(VArrayTarget target, const VArray& a, const VArray& b);