+++ src/vatensor/varray.cpp
@@ -8,11 +8,12 @@
 #include "vassign.h"
 
 va::DType va::VArray::dtype() const {
+    // TODO this is true now, but may not be in the future!
     return static_cast<DType>(store.index());
 }
 
 size_t va::VArray::size() const {
-    return std::visit([](auto&& carray) { return carray.size(); }, to_compute_variant());
+    return std::accumulate(shape.begin(), shape.end(), 1, std::multiplies<>());;
 }
 
 size_t va::VArray::dimension() const {
@@ -42,7 +43,16 @@ va::VArray va::VArray::slice(const xt::xstrided_slice_vector &slices) const {
     }, store);
 }
 
-void va::VArray::fill(const VConstant value) {
+va::VScalar va::VArray::get_scalar(const axes_type &index) const {
+    // xtensor actually checks later, too, but it just pads with 0 rather than throwing.
+    if (index.size() != dimension()) throw std::invalid_argument("invalid dimension for index");
+
+    return std::visit([&index](auto&& carray) -> VScalar {
+        return std::forward<decltype(carray)>(carray)[index];
+    }, to_compute_variant());
+}
+
+void va::VArray::fill(const VScalar value) {
     auto compute_variant = to_compute_variant();
     va::assign(compute_variant, value);
 }
@@ -65,13 +75,13 @@ size_t va::VArray::size_of_array_in_bytes() const {
     }, to_compute_variant());
 }
 
-va::VArray va::from_constant_variant(VConstant constant) {
-    return std::visit([](auto cconstant){
-        return from_constant(cconstant);
-    }, constant);
+va::VArray va::from_scalar_variant(VScalar scalar) {
+    return std::visit([](auto cscalar){
+        return from_scalar(cscalar);
+    }, scalar);
 }
 
-va::VConstant va::dtype_to_variant(const DType dtype) {
+va::VScalar va::dtype_to_variant(const DType dtype) {
     switch (dtype) {
         case DType::Bool:
             return bool();
@@ -100,7 +110,7 @@ va::VConstant va::dtype_to_variant(const DType dtype) {
     }
 }
 
-va::DType va::variant_to_dtype(const VConstant dtype) {
+va::DType va::variant_to_dtype(const VScalar dtype) {
     return static_cast<DType>(dtype.index());
 }
 
@@ -110,8 +120,8 @@ size_t va::size_of_dtype_in_bytes(const DType dtype) {
     }, dtype_to_variant(dtype));
 }
 
-va::VConstant va::constant_to_dtype(const VConstant v, const DType dtype) {
-    return std::visit([](auto v, const auto t) -> va::VConstant {
+va::VScalar va::scalar_to_dtype(const VScalar v, const DType dtype) {
+    return std::visit([](auto v, const auto t) -> va::VScalar {
         using T = std::decay_t<decltype(t)>;
         return static_cast<T>(v);
     }, v, dtype_to_variant(dtype));
@@ -128,8 +138,8 @@ va::DType va::dtype_common_type(const DType a, const DType b) {
     );
 }
 
-va::VConstant va::VArray::to_single_value() const {
-    return std::visit([](const auto& carray) -> va::VConstant {
+va::VScalar va::VArray::to_single_value() const {
+    return std::visit([](const auto& carray) -> va::VScalar {
         if (carray.size() != 1) {
             throw std::runtime_error("Expected a single element after slicing.");
         }
@@ -140,14 +150,14 @@ va::VConstant va::VArray::to_single_value() const {
     }, to_compute_variant());
 }
 
-va::VArray::operator bool() const { return va::constant_to_type<bool>(to_single_value()); }
-va::VArray::operator int64_t() const { return va::constant_to_type<int64_t>(to_single_value()); }
-va::VArray::operator int32_t() const { return va::constant_to_type<int32_t>(to_single_value()); }
-va::VArray::operator int16_t() const { return va::constant_to_type<int16_t>(to_single_value()); }
-va::VArray::operator int8_t() const { return va::constant_to_type<int8_t>(to_single_value()); }
-va::VArray::operator uint64_t() const { return va::constant_to_type<uint64_t>(to_single_value()); }
-va::VArray::operator uint32_t() const { return va::constant_to_type<uint32_t>(to_single_value()); }
-va::VArray::operator uint16_t() const { return va::constant_to_type<uint16_t>(to_single_value()); }
-va::VArray::operator uint8_t() const { return va::constant_to_type<uint8_t>(to_single_value()); }
-va::VArray::operator double() const { return va::constant_to_type<double>(to_single_value()); }
-va::VArray::operator float() const { return va::constant_to_type<float>(to_single_value()); }
+va::VArray::operator bool() const { return va::scalar_to_type<bool>(to_single_value()); }
+va::VArray::operator int64_t() const { return va::scalar_to_type<int64_t>(to_single_value()); }
+va::VArray::operator int32_t() const { return va::scalar_to_type<int32_t>(to_single_value()); }
+va::VArray::operator int16_t() const { return va::scalar_to_type<int16_t>(to_single_value()); }
+va::VArray::operator int8_t() const { return va::scalar_to_type<int8_t>(to_single_value()); }
+va::VArray::operator uint64_t() const { return va::scalar_to_type<uint64_t>(to_single_value()); }
+va::VArray::operator uint32_t() const { return va::scalar_to_type<uint32_t>(to_single_value()); }
+va::VArray::operator uint16_t() const { return va::scalar_to_type<uint16_t>(to_single_value()); }
+va::VArray::operator uint8_t() const { return va::scalar_to_type<uint8_t>(to_single_value()); }
+va::VArray::operator double() const { return va::scalar_to_type<double>(to_single_value()); }
+va::VArray::operator float() const { return va::scalar_to_type<float>(to_single_value()); }