+++ src/vatensor/vassign.cpp
@@ -15,20 +15,20 @@ void va::assign_nonoverlapping(ComputeVariant& array, const ArrayVariant& value)
     }, array, value);
 }
 
-void va::assign(ComputeVariant& array, VConstant value) {
+void va::assign(ComputeVariant& array, VScalar value) {
     std::visit([](auto& carray, const auto cvalue) {
         using V = typename std::decay_t<decltype(carray)>::value_type;
         carray.fill(static_cast<V>(cvalue));
     }, array, value);
 }
 
-void va::assign(VArrayTarget target, VConstant value) {
+void va::assign(VArrayTarget target, VScalar value) {
 	std::visit([value](auto target) {
 		if constexpr (std::is_same_v<decltype(target), ComputeVariant*>) {
 			va::assign(*target, value);
 		}
 		else {
-			*target = from_constant_variant(value);
+			*target = from_scalar_variant(value);
 		}
 	}, target);
 }