+++ src/vatensor/vassign.h
@@ -15,9 +15,9 @@ namespace va {
 
     void assign(ComputeVariant& array, const ComputeVariant& value);
     void assign_nonoverlapping(ComputeVariant& array, const ArrayVariant& value);
-    void assign(ComputeVariant& array, VConstant value);
+    void assign(ComputeVariant& array, VScalar value);
 
-    void assign(VArrayTarget target, VConstant value);
+    void assign(VArrayTarget target, VScalar value);
 }
 
 #endif //NUMDOT_VASSIGN_H