+++ docs/setup/changelog.rst
@@ -19,7 +19,7 @@ Upcoming Changes (main branch)
 
 **Changed**
 
-- ``nd.array`` and ``nd.as_array`` are now up to 2x faster.
+- ``nd.array`` and ``nd.as_array``, ``NDArray.get_float``, ``NDArray.get_int``, ``NDArray.get_bool`` are now up to 2x faster.
 - ``NDArray.to_godot_array`` now slices into the outermost dimension instead of flattening the array. To get floats and ints directly, use ``.to_packedxxx``.
 - ``NDArray.to_packed_xxx`` now require 0D or 1D arrays to work. If the array is 2D, the conversion is not trivial, and a reshape should be used first.
 - NumDot now uses ``Vector4i`` as a surrogate for range objects. They are represented as (bitmask, start, stop, step). This optimizes range creation, interpretation and memory use.